refactoring

+++ src/main/java/by/itacademy/service/api/IAdminService.java
@@ -15,5 +15,5 @@ public interface IAdminService {
 
     PageUserDto get(UUID uuid);
 
-    void update(UUID uuid, int dtUpdate, UserCreateDto user);
+    void update(UUID uuid, long dtUpdate, UserCreateDto user);
 }
